% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{example_fetch}
\alias{example_fetch}
\title{Retrieve PartitionFinder example files}
\usage{
example_fetch(type = c("nucleotide", "morphology", "aminoacid"), dr = getwd())
}
\arguments{
\item{type}{Either nucleotide, morphology or aminoacid}

\item{dr}{Directory to bring the files}
}
\description{
Bring PartitionFinder example files for demonstrations of either
nucleotide, morphology or amino acid runs into a directory, default working
directory.
}
\examples{
library(outsider)
PartitionFinder <- module_import('PartitionFinder',
                                 repo = 'dombennett/om..partitionfinder')
example_fetch <- module_import('example_fetch',
                               repo = 'dombennett/om..partitionfinder')
# bring PartitionFinder examples to working directory
example_fetch(type = 'nucleotide')
# run on the "nucleotide" example folder
PartitionFinder(arglist = 'nucleotide')
}
