% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PartitionFinder}
\alias{PartitionFinder}
\title{PartitionFinder 2}
\usage{
PartitionFinder(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run PartitionFinder 2
}
\examples{
library(outsider)
PartitionFinder <- module_import('PartitionFinder',
                                 repo = 'dombennett/om..partitionfinder')
# bring PartitionFinder examples to working directory
example_fetch(type = 'dna')
PartitionFinder('-h')
}
