% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PartitionFinder}
\alias{PartitionFinder}
\title{PartitionFinder 2}
\usage{
PartitionFinder(arglist = arglist_get(...), outdir = getwd())
}
\arguments{
\item{arglist}{Arguments for PartitionFinder}

\item{outdir}{Filepath to output files are placed}
}
\description{
Run PartitionFinder 2
}
\examples{
library(outsider)
PartitionFinder <- module_import('PartitionFinder',
                                 repo = 'dombennett/om..partitionfinder')
example_fetch <- module_import('example_fetch',
                               repo = 'dombennett/om..partitionfinder')
# bring PartitionFinder examples to working directory
example_fetch(type = 'nucleotide')
# run on the "nucleotide" example folder
PartitionFinder('nucleotide')
}
