% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PartitionFinderMorphology}
\alias{PartitionFinderMorphology}
\title{PartitionFinderMorphology 2}
\usage{
PartitionFinderMorphology(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run PartitionFinderMorphology 2
}
\examples{
library(outsider)
PartitionFinderMorphology <- module_import('PartitionFinderMorphology',
                                 repo = 'dombennett/om..partitionfinder')
example_fetch <- module_import('example_fetch',
                               repo = 'dombennett/om..partitionfinder')
# bring PartitionFinder examples to working directory
example_fetch(type = 'morphology')
# run on the "morphology" example folder
PartitionFinderMorphology('morphology')
}
